+++ packages/atomic/src/components/result-lists/atomic-result-children/atomic-result-children.tsx
@@ -157,6 +157,7 @@ export class AtomicResultChildren {
         content={content}
         result={child}
         engine={this.bindings.engine}
+        store={this.bindings.store}
         density={this.displayConfig.density}
         imageSize={this.imageSize || this.displayConfig.imageSize}
         classes={`child-result ${isLast ? 'last-child' : ''}`.trim()}